Add the glue to allow the RepositoryItem to construct a BagItLibrary Metadata object
This commit fixes the bugs which weren't fixed in the last commit.  The main issue was the xom does not build from a String - when you pass it a string, it expects this to be a URI to a remote document.  This is easily fixed with a ByteArrayInputStream.  Further to this, there are a few more workarounds that have had to be implemented to translate from an Abdera Element to a XOM Element

One day, maybe XML libraries will be nicer to use.